use instanceof instead of calling virtual function GetType()

+++ js/Pixel/Magma.ts
@@ -14,24 +14,22 @@ class MagmaPixel extends WaterPixel {
     for (let y = -1; y <= 1; y++) {
       for (let x = -1; x <= 1; x++) {
         const pixel = world.GetPixel({  x: this.GetPosition().x + x, y: this.GetPosition().y + y });
-        if (pixel !== null) {
-          if (pixel.GetType() === PixelType.Oil) {
-            world.DeletePixel(this.GetPosition());
-            world.Explode(pixel.GetPosition(), 30, 60);
-          } 
-          else if (pixel.GetType() === PixelType.Fuse) {
-            Pixel.FloodFillSearch(pixel.GetPosition(), world, (location) => {
-              world.Explode(location, 30, 60);
-              return false;
-            });
-          }
-          else if (world.DestroyPixel(pixel.GetPosition(), this.GetDurability())) {
-            didDestroy = true;
-          }
+        if (pixel instanceof OilPixel) {
+          world.DeletePixel(this.GetPosition());
+          world.Explode(pixel.GetPosition(), 30, 60);
+        } 
+        else if (pixel instanceof FusePixel) {
+          Pixel.FloodFillSearch(pixel.GetPosition(), world, (location) => {
+            world.Explode(location, 30, 60);
+            return false;
+          });
+        }
+        else if (world.DestroyPixel(pixel.GetPosition(), this.GetDurability())) {
+          didDestroy = true;
         }
       }
     }
-    if (didDestroy) world.DestroyPixel(this.GetPosition(), Number.MAX_SAFE_INTEGER);
+    if (didDestroy) world.DeletePixel(this.GetPosition());
   }
   
   public GetDurability(): number {
